% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_Utils.R
\name{matrix_log_sum_exp}
\alias{matrix_log_sum_exp}
\title{Stable Log-Sum-Exp}
\usage{
matrix_log_sum_exp(values)
}
\arguments{
\item{values}{Numeric vector.}
}
\value{
Numeric scalar.
}
\description{
Stable Log-Sum-Exp
}
\keyword{internal}
