% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_Noise_BR.R
\name{matrix_noise_br_log_density}
\alias{matrix_noise_br_log_density}
\title{Log Density for BR Matrix Noise}
\usage{
matrix_noise_br_log_density(x_list, support)
}
\arguments{
\item{x_list}{List of matrices to evaluate.}

\item{support}{Convex-hull support from \code{matrix_noise_convex_hull_support()}.}
}
\value{
Numeric vector of log-densities.
}
\description{
Log Density for BR Matrix Noise
}
\keyword{internal}
