% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_Utils.R
\name{matrix_validate_x_list}
\alias{matrix_validate_x_list}
\title{Validate Matrix List Input}
\usage{
matrix_validate_x_list(x_list)
}
\arguments{
\item{x_list}{List of matrices to validate.}
}
\value{
A list of same-sized matrices.
}
\description{
Validate Matrix List Input
}
\keyword{internal}
