% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_Init.R
\name{matrix_mixture_kmeans_init}
\alias{matrix_mixture_kmeans_init}
\title{K-Means Initialization for Matrix Mixture Models}
\usage{
matrix_mixture_kmeans_init(x_list, g, nstart = 10)
}
\arguments{
\item{x_list}{A list of numeric matrices, each of dimension r × p}

\item{g}{Integer: number of mixture components}

\item{nstart}{Integer: number of k-means restarts (default: 10)}
}
\value{
A list containing initial parameters.
}
\description{
K-Means Initialization for Matrix Mixture Models
}
\keyword{internal}
